src/requirements.txt
azure-ai-projects==1.0.0b10
azure-identity==1.23.0
